Apply opensafely-core/repo-template
We apply the configuration found in the opensafely-core/repo-template
repo to this repo and replace the shell scripts in the bin directory
with a justfile. We make a small number of changes to the configuration,
most importantly setting the Python version to 3.8, which is that used
by opensafely-core/python-docker.

We upgrade dev and prod dependencies to ensure that everything looks
right (e.g. requirements.dev.txt and requirements.prod.txt contain
hashes). However, we expect to downgrade the prod dependencies, to
reflect opensafely-core/python-docker.

We remove some unnecessary files, such as .python-version and setup.cfg,
that we thought were necessary before we designed and implemented
reusable actions.