Rename "level_1" to "deciles"
Previously, it wasn't necessary to rename "level_1", which is created by
the call to .quantile, to "deciles" because this method used the name of
DECILES (i.e.  Series.name). However, it now is, which may be a
regression in Pandas. When we downgrade Pandas (a prod dependency) to
match opensafely-core/python-docker, we should revisit.