Drop zero-denominator rows by testing value for inf
It's non-trivial to identify the denominator column without the
associated Measure instance. It's much easier to test the value column
for inf, which is returned by Pandas when the second argument of a
division operation is zero.

If we test value for inf, then we can also remove the _get_denominator
helper function, which reduces our use of DataFrame.attrs.

Fixes #17