Log number of zero-denominator rows dropped
We replicate cohort-extractor's logging using the logging package,
rather than the structlog package, because the latter is a transitive
dependency of python-docker.

I've called `logger.info` in preference to `print`, because when reading
a log file, my eye skips over lines that don't look like log entries.

My intention was to log the number of zero-denominator rows dropped. I
realised that `mask` wasn't a good name (what is it masking?), so I
renamed it to `is_not_inf`.