Expose `show_outer_percentiles` arg
To expose this arg, we add the mechanism for reading JSON-encoded
configuration, which we expect to be passed by job-runner. As such, we
don't check that the JSON is valid. However, we do check that the
configuration keys are correct, and report any incorrect configuration
keys (e.g. typos) to the user. We don't check the configuration values
are correct; doing so would require more complex code, which is probably
unnecessary.

We should still consider what we'd like the `config` key to look like.
However, we can iterate on that problem as we add more configuration.

Closes #21