Tag new versions sequentially
By moving the tag-new-version job into a new workflow, and making this
workflow dependant on the CI workflow, we ensure that new versions are
tagged sequentially. Doing so removes the risk of a race condition
tagging new versions in an incorrect order.

As a bonus, we also run the CI workflow more frequently: not only on
push to the main branch, but also following several other events. Doing
so ensures that new versions are correct, insofar as the check, test,
and test-integration jobs are concerned.

Fixes #113